+++ cmd/go-mysqlbinlog/main.go
@@ -49,7 +49,7 @@ func main() {
 
 	b := replication.NewBinlogSyncer(cfg)
 
-	pos := mysql.Position{Name: *file, Pos: uint32(*pos)}
+	pos := mysql.Position{Name: *file, Pos: uint64(*pos)}
 	if len(*backupPath) > 0 {
 		// Backup will always use RawMode.
 		err := b.StartBackup(*backupPath, pos, 0)
