+++ failover/server.go
@@ -152,7 +152,7 @@ func (s *Server) FetchSlaveReadPos() (Position, error) {
 	fname, _ := r.GetStringByName(0, "Master_Log_File")
 	pos, _ := r.GetIntByName(0, "Read_Master_Log_Pos")
 
-	return Position{Name: fname, Pos: uint32(pos)}, nil
+	return Position{Name: fname, Pos: uint64(pos)}, nil
 }
 
 // FetchSlaveExecutePos gets current executed binlog filename and position from master
@@ -165,7 +165,7 @@ func (s *Server) FetchSlaveExecutePos() (Position, error) {
 	fname, _ := r.GetStringByName(0, "Relay_Master_Log_File")
 	pos, _ := r.GetIntByName(0, "Exec_Master_Log_Pos")
 
-	return Position{Name: fname, Pos: uint32(pos)}, nil
+	return Position{Name: fname, Pos: uint64(pos)}, nil
 }
 
 func (s *Server) MasterPosWait(pos Position, timeout int) error {
